+++ tests/fields/test_BaseField.py
@@ -2,35 +2,33 @@ from dataclasses_avroschema import fields
 
 
 def test_render():
-    field = fields.Field("first_name", str, fields.NULL, metadata={"desc": "English Language Name"})
+    field = fields.Field("first_name", str, metadata={"desc": "English Language Name"})
 
     expected = {
         "name": "first_name",
-        "type": ["string", fields.NULL],
-        "default": fields.NULL,
+        "type": "string",
         "desc": "English Language Name",
     }
 
     assert expected == field.render()
 
-    field = fields.Field("engine_name", str, fields.NULL)
+    field = fields.Field("engine_name", str)
 
     expected = {
         "name": "engine_name",
-        "type": ["string", fields.NULL],
-        "default": fields.NULL,
+        "type": "string",
     }
 
     assert expected == field.render()
 
     field = fields.Field(
-        "breed_name", str, fields.NULL, metadata={"encoding": "some_exotic_encoding", "doc": "Official Breed Name"},
+        "breed_name", str, "test", metadata={"encoding": "some_exotic_encoding", "doc": "Official Breed Name"},
     )
 
     expected = {
         "name": "breed_name",
-        "type": ["string", fields.NULL],
-        "default": fields.NULL,
+        "type": "string",
+        "default": "test",
         "encoding": "some_exotic_encoding",
         "doc": "Official Breed Name",
     }
@@ -39,20 +37,20 @@ def test_render():
 
 
 def test_render_metadata():
-    field = fields.Field("first_name", str, fields.NULL, metadata={"desc": "English Language Name"})
+    field = fields.Field("first_name", str, metadata={"desc": "English Language Name"})
 
     expected = [("desc", "English Language Name")]
 
     assert expected == field.get_metadata()
 
-    field = fields.Field("engine_name", str, fields.NULL)
+    field = fields.Field("engine_name", str)
 
     expected = []
 
     assert expected == field.get_metadata()
 
     field = fields.Field(
-        "breed_name", str, fields.NULL, metadata={"encoding": "some_exotic_encoding", "doc": "Official Breed Name"},
+        "breed_name", str, metadata={"encoding": "some_exotic_encoding", "doc": "Official Breed Name"},
     )
 
     expected = [("encoding", "some_exotic_encoding"), ("doc", "Official Breed Name")]